Replace sim names with more easily understood names.
I'm happy to change these names to whatever people like, but it seemed very
difficult to have to look up what each sim was every time it was referenced in
the text. I also tried to remove references to "baseline PHA cadence" since it
seems easily confused with the real "baseline" cadence.